Refactored ADC conversion routine. No longer a state machine where each state primes the next; PIT0 periodically checks to see if a new conversion is due, primes it and signals the ADC0 IRQ with the correct instruction. Priority is given to the next conversion in a round-robin manner.

+++ FRDM-TFC/Project_Headers/telemetry/Collector.h
@@ -18,9 +18,9 @@ typedef struct {
 	uint8_t   deref; /* Number of deferences required to reach data -
 	                   facilitates access to data behind pointers */
 	Endpoint  endpoint;
-	uint16_t  frequency;
-	uint16_t  period;
-	uint16_t  counter;
+	uint32_t  frequency;
+	uint32_t  period;
+	uint32_t  counter;
 } CollectorItem;
 
 /* Endpoint error reporting */
